backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-stubs>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
]
